bench-pthread-create: Thread creation benchmark
This commit adds a new benchmark to measure the time to create a new
thread. We use C++ std::thread so we have some additional overhead to
pthread_create(). Furthermore, we measure the time until a newly
createed thread runs for the first time, so this is actually measuring
the time to create a new thread and its scheduling delay.