benchmark: Add "no interference" scenario
Add a scenario for running a benchmark without any inteference (a single
threaded benchmark) for those benchmarks where it's possible. For
example, the eventfd benchmark requires two threads and cannot run in
this mode, but most of the benchmarks can be. The "no inteference"
scenario shows the baseline performance of the benchmarked operation.