bench-mmap/munmap: run operations more
The benchmarks complete under 1 second, which does not capture the full
effect of deferred work, for example. Run the benchmarks for bit longer
to address that. In the future, we ought to switch the harness to run a
fixed time slice (e.g. 5 seconds) instead of relying on some operation
count, which causes time to vary between benchmarks.