bench-pthread-kill: Fix race that causes lost wakeup
Fix a race that causes pthread_cond_wait() wakeup to be lost. The scenario is
as follows: T1 (the measuring thread) sends a signal to T2 (the interfering
thread), but T2 manages to handle the signal with sigtimedwait() and
race to pthread_cond_signal() before T1 does pthread_cond_wait(). Fix
the issue by sending the signal in T1 under the serializing mutex, which
ensures that pthread_cond_wait() is always called *before*
pthread_cond_signal() even if signal is handled fast by T2.